fix: watch buffers instead of editors
Watching `TextBuffer` objects instead of `TextEditor` objects works
better, because editors of the same file share the same buffer.
Also, the `onDidDestroy` listener was broken, because it didn't
look for editors of the same path before mutating the `watching`
object. The first solution I thought of was to keep track of how
many editors are open per watched file, but that also meant
attaching `onDidDestroy` listeners to each editor. I found it much
simpler to attach listeners to the shared TextBuffer objects.

The package of each watched file has an `onDidDeactivate`
listener attached. This allows us to stop watching the files
of that package when/if the grammar is deactivated. This listener
is reattached whenever the grammar is reloaded, because the `Package`
object is recreated.

When a watched file is deleted, we stop watching it.

When a watched file is renamed, we stop watching it *if* the new file
path does not match the `grammarRE` pattern. We assume its package
is still the same, but maybe this isn't a safe assumption?

We *don't* yet listen for renamed package of watched files, but
this could be a nice addition sometime in the future.

The buffer listeners are now properly disposed of when live reloading
is disabled by the user.

The `grammarRE` string now has its '/' characters properly escaped.